Improve Config class
- The goal is to reduce cognitive learning and to improve
code maintainability. The following was done:

- The Config class is now Immutable. Every change made to the config
create a new config object instead of updating the currently used.

- The Array indexes are no longer abbreviated as the method names too.

- To take advantage of TypeHint the Formatter and Cache setter no longer
takes anything other than an object which implents the corresponding interface

- Some settings validation have also been improved.

- Since The cache setter method no longer accept boolean value,
the NullCache object which implements the CacheInterface was added to enabled the
use of no cache.

Since the Config class is now an Immutable Value Object, the Database object was
also updated to take into account this change. Moreover the public API from the
Database class was reduced to only exposed the used methods.

The test suite was also modified according to the changes.